Delay tf.cast on Shepard Goodness

+++ src/tensorflow_projection_qm/metrics/shepard_goodness.py
@@ -10,13 +10,13 @@ def shepard_goodness_impl(X, X_2d):
     flat_d_low = distance.flat_psqdist(X_2d)
     n = tf.cast(tf.shape(flat_d_high)[0], tf.float32)
 
-    ranks_high = tf.cast(tf.argsort(tf.argsort(flat_d_high)), tf.float32)
-    ranks_low = tf.cast(tf.argsort(tf.argsort(flat_d_low)), tf.float32)
+    ranks_high = tf.argsort(tf.argsort(flat_d_high))
+    ranks_low = tf.argsort(tf.argsort(flat_d_low))
 
     # Assuming no ties, we can use this formula.
-    return 1 - 6 * tf.reduce_sum(tf.math.squared_difference(ranks_high, ranks_low)) / (
-        n * (n**2 - 1)
-    )
+    return 1 - 6 * tf.cast(
+        tf.reduce_sum(tf.math.squared_difference(ranks_high, ranks_low)), tf.float32
+    ) / (n * (n**2 - 1))
 
 
 def shepard_goodness(X, X_2d):
